collect the error from AssistMCU and info DetectMCU.
1/ DetectMCU add the ERROR_Status and related handle.
2/ mainboard add Inform_Detect function, and related CMD code
in the check_ready function,also check error.and handle it.

notice:the codes from mainboard to DetectMcu are different from the codes from Assist to mainboard.
the code to send or check need to be neaten,communication has too many patches.
In mainboard,the check of error is in sensor module,and the handle is also in module. it is somehow comfusing.